# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Progetto Gruppo T31 - IS2022",
        "description": "API del progetto del Gruppo T31 svolto durente il corso di Ingegneria del software - Anno Accademico 2022/23",
        "contact": {
            "name": "Luca Dematté - Referente gruppo T31",
            "email": "luca.dematte-2@studenti.unitn.it"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "gruppo-t31-api.onrender.com",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "paths": {
...
        "/spazio": {
            "post": {
                "tags": ["spazio"],
                "summary": "Creazione di un nuovo spazio",
                "description": "",
                "operationId": "nuovoSpazio",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "nome",
                        "in": "formData",
                        "description": "Nome del nuovo spazio",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "descrizione",
                        "in": "formData",
                        "description": "Breve descrizione del nuovo spazio",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "tipologia",
                        "in": "formData",
                        "description": "Tipo di spazio",
                        "type": "string",
                        "enum": ["Locale interno", "Locale attrezzato", "Spazio esterno"],
                        "required": true
                    },
                    {
                        "name": "prezzoIniziale",
                        "in": "formData",
                        "description": "Prezzo dello spazio quando viene utilizzato",
                        "type": "number",
                        "format": "float",
                        "required": true
                    },
                    {
                        "name": "prezzoOra",
                        "in": "formData",
                        "description": "Prezzo dello spazio per ogni ora di utilizzo",
                        "type": "number",
                        "format": "float",
                        "required": true
                    },
                    {
                        "name": "URLfoto",
                        "in": "formData",
                        "description": "Percorso della foto del nuovo spazio",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "servizi",
                        "in": "formData",
                        "description": "Lista di ID dei servizi che possono essere utilizzati in questo spazio",
                        "type": "array",
                        "items": { "type": "string" },
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created - Creazione spazio effettuata correttamente",
                        "schema": { "$ref": "#/definitions/Spazio" }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            },
            "get": {
                "tags": ["spazio"],
                "summary": "Ottenere la lista degli spazi presenti nel sistema",
                "description": "",
                "operationId": "listaSpazi",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Indice del primo oggetto da restituire",
                        "type": "integer",
                        "required": false
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Numero di oggetti da restituire",
                        "type": "integer",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Spazio" }
                        }
                    },
                    "204": {
                        "description": "No content - Nessuno spazio da restituire",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer",
                                    "description": "Codice di risposta HTTP",
                                    "example": 204
                                },
                                "message": {
                                    "type": "string",
                                    "description": "Messaggio con esito dell'operazione",
                                    "example": "Nessuno spazio da restituire"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - I valori di start e count non sono coerenti",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/spazio/byID": {
            "get": {
                "tags": ["spazio"],
                "summary": "Ottenere uno spazio tramite il suo ID",
                "description": "",
                "operationId": "getSpazioConID",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID dello spazio da restituire",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/Spazio" }
                    },
                    "404": {
                        "description": "Not found - Uno spazio con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
...
        "/spazio/{id}": {
            "patch": {
                "tags": ["spazio"],
                "summary": "Modifica di uno spazio",
                "description": "",
                "operationId": "modificaSpazio",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dello spazio da modificare",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "nome",
                        "in": "formData",
                        "description": "Nome dello spazio",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "descrizione",
                        "in": "formData",
                        "description": "Breve descrizione dello spazio",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "tipologia",
                        "in": "formData",
                        "description": "Tipo di spazio",
                        "type": "string",
                        "enum": ["Locale interno", "Locale attrezzato", "Spazio esterno"],
                        "required": true
                    },
                    {
                        "name": "prezzoIniziale",
                        "in": "formData",
                        "description": "Prezzo dello spazio quando viene utilizzato",
                        "type": "number",
                        "format": "float",
                        "required": true
                    },
                    {
                        "name": "prezzoOra",
                        "in": "formData",
                        "description": "Prezzo dello spazio per ogni ora di utilizzo",
                        "type": "number",
                        "format": "float",
                        "required": true
                    },
                    {
                        "name": "URLfoto",
                        "in": "formData",
                        "description": "Percorso della foto dello spazio",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "servizi",
                        "in": "formData",
                        "description": "Lista di ID dei servizi che possono essere utilizzati in questo spazio",
                        "type": "array",
                        "items": { "type": "string" },
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Modifica spazio effettuata correttamente",
                        "schema": { "$ref": "#/definitions/Spazio" }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Uno spazio con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            },
            "delete": {
                "tags": ["spazio"],
                "summary": "Eliminazione di uno spazio",
                "description": "",
                "operationId": "eliminaSpazio",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID dello spazio da eliminare",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Eliminazione dello spazio effettuata correttamente",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer",
                                    "description": "Codice di risposta HTTP",
                                    "example": 200
                                },
                                "message": {
                                    "type": "string",
                                    "description": "Messaggio con esito dell'operazione",
                                    "example": "Spazio eliminato correttamente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Uno spazio con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Spazio": {
            "type": "object",
            "required": ["_id", "nome", "tipologia", "prezzoIniziale", "prezzoOra"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco dello spazio"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome dello spazio",
                    "example": "Sala polifunzionale"
                },
                "descrizione": {
                    "type": "string",
                    "description": "Breve descrizione dello spazio",
                    "example": "Aula molto grande (100 posti) per spettacoli, conferenze, ..."
                },
                "tipologia": {
                    "type": "string",
                    "description": "Tipo di spazio",
                    "enum": ["Locale interno", "Locale attrezzato", "Spazio esterno"],
                    "example": "Locale interno"
                },
                "prezzoIniziale": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo dello spazio quando viene utilizzato",
                    "example": 10.5
                },
                "prezzoOra": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo dello spazio per ogni ora di utilizzo",
                    "example": 4.25
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto dello spazio",
                    "example": "/images/spazi/default.png"
                },
                "servizi": {
                    "type": "array",
                    "description": "Lista dei servizi che possono essere utilizzati in questo spazio",
                    "items": { "$ref": "#/definitions/Servizio" },
                    "required": false
                }
            }
        },
        "Servizio": {
            "type": "object",
            "required": ["_id", "nome", "tipologia", "prezzoIniziale", "prezzoOra"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco del servizio"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome del servizio",
                    "example": "Proiettore"
                },
                "descrizione": {
                    "type": "string",
                    "description": "Breve descrizione del servizio",
                    "example": "Proiettore 720p con interfaccia HDMI e VGA"
                },
                "tipologia": {
                    "type": "string",
                    "description": "Tipo di servizio",
                    "enum": ["Arredamento", "Attrezzatura specifica", "Altro"],
                    "example": "Attrezzatura specifica"
                },
                "prezzoIniziale": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo del servizio quando viene utilizzato",
                    "example": 2.5
                },
                "prezzoOra": {
                    "type": "number",
                    "format": "float",
                    "description": "Prezzo del servizio per ogni ora di utilizzo",
                    "example": 1.0
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto del servizio",
                    "example": "/images/servizi/default.png"
                }
            }
        },
...
        "ErroreAPI": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
                "code": {
                    "type": "integer",
                    "description": "Codice di risposta HTTP"
                },
                "message": {
                    "type": "string",
                    "description": "Messaggio di testo sull'errore"
                }
            }
        }
    }
}